feat: replace *float64 with custom Decimal type for FHIR decimal precision preservation
FHIR decimal type requires exact representation (e.g. "1.50" must preserve
trailing zeros). The previous *float64 mapping violated this SHALL requirement.

Introduces a Decimal type backed by string internally, preserving exact textual
representation through JSON and XML round-trips. Updates codegen to map
decimal fields to *Decimal across all r4, r4b, and r5 packages.

Closes #6

+++ internal/codegen/analyzer/analyzer_test.go
@@ -223,7 +223,7 @@ func TestFHIRToGoType(t *testing.T) {
 		{"boolean", "bool"},
 		{"integer", "int"},
 		{"integer64", "int64"},
-		{"decimal", "float64"},
+		{"decimal", "Decimal"},
 		{"string", "string"},
 		{"uri", "string"},
 		{"date", "string"},
